requirement.txt
aiohttp==3.8.4
bokeh==2.4.3
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cloudpickle==2.2.0
colorcet==3.0.1
cycler==0.11.0
dataclasses==0.6
datasets==2.8.0
entrypoints==0.4
faiss-cpu==1.7.3
fastapi==0.92.0
fastapi-socketio==0.0.10
fastjsonschema==2.16.3
frozenlist==1.3.3
httptools==0.5.0
huggingface-hub==0.11.1
iopath==0.1.10
ipywidgets==8.0.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lightning-utilities==0.6.0.post0
llvmlite==0.39.1
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
meerkat-ml==0.4.6
murmurhash==1.0.9
nicegui==1.2.1
nltk==3.8.1
numba==0.56.4
numpy==1.23.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
portalocker==2.6.0
prompt-toolkit==3.0.38
pyarrow==11.0.0
python-dateutil==2.8.2
python-multipart==0.0.6
python-socketio==5.8.0
pytorch-lightning==1.9.2
pytz==2022.7.1
pyviz-comms==2.2.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rich==13.3.2
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
# sentencepiece==0.1.94
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
SQLAlchemy==1.4.46
tifffile==2022.10.10
tokenizers==0.13.2
torch==1.13.1
# torchmetrics==0.11.1
# torchvision==0.14.1
transformers==4.25.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.20.0
vbuild==0.8.1
websockets==10.4
widgetsnbextension==4.0.5
xarray==2022.12.0
yarl==1.8.2
zipp==3.11.0
